refactor(explorer): modularize refresh scheduling
